requirements.txt
pandas
python-dotenv
requests
dvc
dvc_gdrive
matplotlib
seaborn
numpy
missingno
requests==2.28.0